Version bump and upgrade file fix
